refactor: LocalState のパス組み立てと読み出しを 1 箇所へ寄せる (#976)
#957 で RecordBgDiagnostic の実体は共有化したのに、パス組み立てと
ReadFileUtf8 は push_background_task.cpp / push_diagnostics_store.cpp /
wns_push.cpp の 3 ファイルに写しが残っていた（ConsumePushDiagnosticsJson を
数えると 4 箇所）。local_state_files.h は「writer と reader が別リテラルで
持つと片方だけ改名したときに黙って壊れる」を避けるためのヘッダ (#764) な
ので、その周りの手続きが割れているのでは意味が薄い。

local_state_files.cpp を足し、LocalStateFilePath / ReadFileUtf8 /
ReadLocalStateFileUtf8 を置いた。runner と capsicum_push_task の両方の
ターゲットへ追加している（push_diagnostics_store.cpp と同じ扱い）。

⚠ LocalStateFilePath は投げない契約にした。非パッケージ起動では LocalFolder
を引けないので空文字列を返し、呼び出し側は「この経路は使えない」と読んで
素通りする。これまで呼び出しごとに try を書くか書かないかが割れていた。

⚠ **この差分は macOS ではコンパイルできない**（WinRT）。実機で
`flutter build windows --debug` を通すまでビルド検証は済んでいない
（docs/tech-notes.md「Windows ネイティブを触ったときの検証手順」）。
